fix(worker): resolve ERR_MODULE_NOT_FOUND in worker container
The worker container's crux.mjs fails to load because it value-imports
constants from wiki-server route files that aren't copied into the image.
Move the shared constants (VALID_SUGGESTION_STATUSES,
TABLEBASE_NON_ROUTE_FILES, VALID_RISK_DOMAINS) into
apps/wiki-server/src/api-types.ts — the one wiki-server file the worker
image already copies (per existing comment in Dockerfile.worker).

Also:
- Add COPY for apps/web/src/data/entity-type-names.ts (value-imported by
  crux/lib/research-playbooks.ts, crux/lib/rules/frontmatter-schema.ts,
  and crux/wiki-server/sync-things.ts).
- Lazy-load archive-pdfs in crux/commands/resources.ts to keep the AWS
  SDK out of the worker image — the worker never runs archive-pdfs.
- Split value/type imports in crux/third-party-evals/extract-eval-report.ts.

This is what caused the backfill-sources cron at 2026-05-18 02:00 UTC to
fail with "Cannot find module .../url-suggestions.ts". Same class of
failure that QUA-605 and QUA-598 already fixed for other modules.
